requirements/base.txt
Django==5.1.4
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
django-cors-headers==3.14.0
drf-yasg==1.21.5
argon2-cffi
redis==4.5.1
celery==5.2.7
flower==1.2.0
django-celery-email==3.0.0
django-allauth==0.52.0
djangorestframework-simplejwt==5.2.2
dj-rest-auth==3.0.0
djangorestframework-gis